requirements.txt
aws-cdk-lib==2.15.0
constructs>=10.0.0,<11.0.0
cdk-ecr-deployment==2.1.3
golang==0.0.1
rice==0.4